Start with image page generation
Of course I would never generate an HTML page like that, but my goal was
to write some Clojure every day in order to gain exposure to the
language and get used to programming again. It was not digging around in
documentation. The next step will of course be doing all that with
Enlive and proper HTML5.